COmplete version of the notebook
